refactor: move all sequencer metrics stuff to one module
