+++ crates/pathfinder/src/sequencer.rs
@@ -1,6 +1,7 @@
 //! StarkNet L2 sequencer client.
 mod builder;
 pub mod error;
+mod metrics;
 pub mod reply;
 pub mod request;
 
@@ -127,7 +128,7 @@ impl Client {
 
     /// Create a Sequencer client for the given [Url].
     pub fn with_url(url: Url) -> reqwest::Result<Self> {
-        Self::register_metrics();
+        metrics::register();
 
         Ok(Self {
             inner: reqwest::Client::builder()
@@ -161,40 +162,6 @@ impl Client {
             integration if integration == INTEGRATION_GENESIS_HASH => Ok(Chain::Integration),
             other => Err(anyhow::anyhow!("Unknown genesis block hash: {}", other.0)),
         }
-    }
-
-    const METRIC_REQUESTS: &'static str = "sequencer_requests_total";
-    const METRIC_FAILED_REQUESTS: &'static str = "sequencer_requests_failed_total";
-    const METRIC_STARKNET_ERRORS: &'static str = "sequencer_requests_failed_starknet_total";
-    const METRIC_DECODE_ERRORS: &'static str = "sequencer_requests_failed_decode_total";
-    const METRIC_RATE_LIMITED: &'static str = "sequencer_requests_failed_rate_limited_total";
-    const METRICS: &'static [&'static str] = &[
-        Self::METRIC_REQUESTS,
-        Self::METRIC_FAILED_REQUESTS,
-        Self::METRIC_STARKNET_ERRORS,
-        Self::METRIC_DECODE_ERRORS,
-        Self::METRIC_RATE_LIMITED,
-    ];
-
-    fn register_metrics() {
-        // We also track `get_block`, `get_state_update` wrt `latest` and `pending` blocks
-        let methods = ["get_block", "get_state_update"].into_iter();
-        let tags = ["latest", "pending"].into_iter();
-
-        // Register counters for all the methods
-        Self::METRICS.iter().for_each(|&name| {
-            builder::Request::<'_, builder::stage::Method>::METHODS
-                .iter()
-                .for_each(|&method| {
-                    metrics::register_counter!(name, "method" => method);
-                });
-
-            methods.clone().for_each(|method| {
-                tags.clone().for_each(|tag| {
-                    metrics::register_counter!(name, "method" => method, "tag" => tag);
-                })
-            })
-        });
     }
 }
